% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{get_info}
\alias{get_info}
\title{Get information about oak and truffles}
\usage{
get_info(dboak, dbtruffle, dbreensemence, theidoak)
}
\arguments{
\item{dboak}{Database containing information about oak trees.}

\item{dbtruffle}{Database containing information about truffles.}

\item{dbreensemence}{Database containing reseeding information.}

\item{theidoak}{ID of the oak tree for which information is to be retrieved.}
}
\value{
A list containing information about oak trees and truffles.
}
\description{
This function retrieves information about oak trees and truffles based on the provided
databases and oak tree ID.
}
\examples{
conn <- DBI::dbConnect(RSQLite::SQLite(), system.file("chenes_truffe.sqlite", package = "truffles"))
chene <- DBI::dbReadTable(conn, name = "chenes")
truffe <- DBI::dbReadTable(conn, name = "truffe")
reensemence <- DBI::dbReadTable(conn, name = "reens")

get_info(dboak = chene, dbtruffle = truffe, dbreensemence = reensemence, theidoak = "162")
DBI::dbDisconnect(conn)
}
