% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_chene_last_truffe.R
\name{get_info_chene_last_truffe}
\alias{get_info_chene_last_truffe}
\title{Get Information about the Last Truffle Found on a Specific Oak Tree}
\usage{
get_info_chene_last_truffe(dbtruffe, theidchene)
}
\arguments{
\item{dbtruffe}{A data frame containing information about truffles.}

\item{theidchene}{The ID of the oak tree.}
}
\value{
A data frame containing information about the last truffle found on the specified oak tree.
}
\description{
This function retrieves information about the last truffle found on a specific oak tree.
}
\examples{

dbtruffe <- data.frame(
  idchene = c(123, 123, 456, 789),
  date_trouve = as.Date(c("2023-01-01", "2023-03-15", "2023-02-01", "2022-12-01"))
)
get_info_chene_last_truffe(dbtruffe = dbtruffe, theidchene = 123)
}
