% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_chene_truffe.R
\name{get_info_chene_truffe}
\alias{get_info_chene_truffe}
\title{get_info_chene_truffe Title}
\usage{
get_info_chene_truffe(dbtruffe, theidchene)
}
\arguments{
\item{dbtruffe}{dbtruffe TOCOMPLETE}

\item{theidchene}{idchene TOCOMPLETE}
}
\value{
a list
}
\description{
get_info_chene_truffe Title
}
\examples{
conn <- connect_db()
truffe <- DBI::dbReadTable(conn, name = "truffe")

get_info_chene_truffe(dbtruffe = truffe, theidchene = "119")
DBI::dbDisconnect(conn)
}
