% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_reensemence.R
\name{get_info_reensemence}
\alias{get_info_reensemence}
\title{Get the latest reseeding date for a given oak tree ID.}
\usage{
get_info_reensemence(dbreensemence, theidchene)
}
\arguments{
\item{dbreensemence}{A data frame containing reseeding information with columns "idchene" and "date_reens".}

\item{theidchene}{The ID of the oak tree for which to retrieve the reseeding date.}
}
\value{
Returns the latest reseeding date for the specified oak tree ID. If no reseeding
information is available for the given ID, returns "-".
}
\description{
This function retrieves the latest reseeding date for a specified oak tree ID
from a dataframe containing reseeding information.
}
\examples{
conn <- connect_db()
reensemence <- DBI::dbReadTable(conn, name = "reens")

get_info_reensemence(dbreensemence = reensemence, theidchene = "150")
DBI::dbDisconnect(conn)
}
