% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_leaflet.R
\name{prepare_leaflet}
\alias{prepare_leaflet}
\title{Prepare Leaflet Data}
\usage{
prepare_leaflet(dbchene, dbtruffe, dbreens)
}
\arguments{
\item{dbchene}{A data frame containing information about oak trees.}

\item{dbtruffe}{A data frame containing information about truffles.}

\item{dbreens}{A data frame containing information about reensemencement.}
}
\value{
A list suitable for passing to Leaflet JavaScript for visualization.
}
\description{
This function prepares data for Leaflet JavaScript visualization.
}
\examples{
dbchene <- data.frame(
  idoak = c(1, 2, 3),
  chene_nom = c("Chene 1", "Chene 2", "Chene 3")
)

dbreens <- data.frame(
  idreens = c(8),
  idoak = c(2),
  date_reens = c("2022-04-30")
)

dbtruffe <- data.frame(
  idoak = c(1, 1, 3),
  idtruffle = c(1, 2, 3),
  estimation = c(1, 0, 1),
  weight = c(10, 15, NA)
)
prepare_leaflet(dbchene, dbtruffe, dbreens)
}
