% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truffles-package.R
\docType{package}
\name{truffles-package}
\alias{truffles}
\alias{truffles-package}
\title{truffles: Truffles}
\description{
To map and manage a truffle tree plantation.
}
\author{
\strong{Maintainer}: Murielle Delmotte \email{murielle@thinkr.fr}

}
\keyword{internal}
