% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truffles_by_year.R
\name{truffles_by_year}
\alias{truffles_by_year}
\title{Title}
\usage{
truffles_by_year(truffles)
}
\arguments{
\item{truffles}{database}
}
\value{
a dataframe
}
\description{
Description
}
\examples{
conn <- connect_db()
truffes <- DBI::dbReadTable(conn, name = "truffe")

truffles_by_year(truffes)
}
