% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_db_truffe.R
\name{update_db_truffe}
\alias{update_db_truffe}
\title{Update a row in the "truffe" table of a database.}
\usage{
update_db_truffe(
  conn = connect_db(),
  idtruffe,
  idchene,
  date_trouve,
  poids,
  commentaires,
  estimation
)
}
\arguments{
\item{conn}{A connection to the database.}

\item{idtruffe}{The identifier of the truffle to update.}

\item{idchene}{The identifier of the oak tree associated with the truffle.}

\item{date_trouve}{The date when the truffle was found.}

\item{poids}{The weight of the truffle.}

\item{commentaires}{The comments associated with the truffle.}

\item{estimation}{The estimation of the truffle.}
}
\value{
The function does not return any value, but it updates the corresponding row in the "truffe" table of the database.
}
\description{
This function updates a row in the "truffe" table of a database using an SQL UPDATE query.
}
\examples{

conn <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")
  
  # Création de la table truffe dans la base de données
  DBI::dbWriteTable(
    conn,
    "truffe",
    data.frame(
      idtruffe = "156",
      idchene = "15",
      date_trouve = as.numeric(as.Date("2020/02/02")),
      poids = 12,
      estimation = 1,
      commentaires = "Comment"
    )
  )

update_db_truffe(conn, idtruffe = "156", idchene = "15", date_trouve = as.numeric(as.Date("2020/02/02")), poids = 22, 
                 estimation = 0, commentaires = "New comment")
}
