% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_db_truffle.R
\name{update_db_truffle}
\alias{update_db_truffle}
\title{Update a row in the "truffe" table of a database.}
\usage{
update_db_truffle(
  conn,
  idtruffle,
  idoak,
  date_found,
  weight,
  comment,
  estimation
)
}
\arguments{
\item{conn}{A connection to the database.}

\item{idtruffle}{The identifier of the truffle to update.}

\item{idoak}{The identifier of the oak tree associated with the truffle.}

\item{date_found}{The date when the truffle was found.}

\item{weight}{The weight of the truffle.}

\item{comment}{The comments associated with the truffle.}

\item{estimation}{The estimation of the truffle.}
}
\value{
The function does not return any value, but it updates the corresponding row in the "truffe" table of the database.
}
\description{
This function updates a row in the "truffe" table of a database using an SQL UPDATE query.
}
\examples{

conn <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")

# Create a truffle table in the database
DBI::dbWriteTable(
  conn,
  "truffe",
  data.frame(
    idtruffle = "156",
    idoak = "15",
    date_found = as.numeric(as.Date("2020/02/02")),
    weight = 12,
    estimation = 1,
    comment = "Comment"
  )
)

update_db_truffle(
  conn,
  idtruffle = "156",
  idoak = "15",
  date_found = as.numeric(as.Date("2020/02/02")),
  weight = 22,
  estimation = 0,
  comment = "It's good!"
)
}
