% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_truffles_by.R
\name{weight_truffles_by}
\alias{weight_truffles_by}
\title{Title}
\usage{
weight_truffles_by(truffles, ...)
}
\arguments{
\item{truffles}{database}

\item{...}{TOCOMPLETE}
}
\value{
a dataframe
}
\description{
Description
}
\examples{

library(dplyr)

conn <- connect_db()
truffes <- DBI::dbReadTable(conn, name = "truffe")

weight_truffles_by(truffes, annee)

truffes_chene <- truffes |> 
  inner_join(DBI::dbReadTable(conn, name = "chenes_feularde"), by = join_by(idchene == id))
weight_truffles_by(truffes_chene, annee, type)
}
