% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_db_new_truffe.R
\name{write_db_new_truffe}
\alias{write_db_new_truffe}
\title{write_db_new_truffe}
\usage{
write_db_new_truffe(
  conn = connect_db(),
  theidchene,
  date_trouvee,
  poids,
  comment,
  digest_ = Sys.time()
)
}
\arguments{
\item{conn}{a connection to database}

\item{theidchene}{TOCOMPLETE}

\item{date_trouvee}{TOCOMPLETE}

\item{poids}{TOCOMPLETE}

\item{comment}{TOCOMPLETE}

\item{digest_}{TOCOMPLETE}
}
\description{
Description
}
\examples{
# write_db_new_truffe(theidchene = "chene")
}
