% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_db.R
\name{connect_db}
\alias{connect_db}
\title{connect_db}
\usage{
connect_db(
  dbname = "postgres",
  host = Sys.getenv("DB_HOST", unset = ""),
  port = 6543,
  user = Sys.getenv("DB_USER", unset = ""),
  password = Sys.getenv("DB_PWD", unset = "")
)
}
\arguments{
\item{dbname}{name of db}

\item{host}{host of db}

\item{port}{port of db}

\item{user}{user of db}

\item{password}{password of db}
}
\description{
To connect to the database
}
\examples{
if (interactive()) {
  connect_db()
}
}
