% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_db.R
\name{connect_db}
\alias{connect_db}
\title{connect_db}
\usage{
connect_db(dbname = "chenes_truffe.sqlite")
}
\arguments{
\item{dbname}{name of db available in inst/}
}
\description{
To connect to the database available in the {truffles} package
}
\examples{
connect_db()
}
