% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{get_info}
\alias{get_info}
\title{Title}
\usage{
get_info(dbchene, dbtruffe, theidchene)
}
\description{
Description
}
\examples{
conn <- DBI::dbConnect(RSQLite::SQLite(), system.file("chenes_truffe.sqlite", package = "truffles"))
chene <- DBI::dbReadTable(conn, name = "chenes_feularde")
truffe <- DBI::dbReadTable(conn, name = "truffe")

get_info(dbchene = chene, dbtruffe = truffe, theidchene = "162")
DBI::dbDisconnect(conn)
}
