% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_chene.R
\name{get_info_chene}
\alias{get_info_chene}
\title{Title}
\usage{
get_info_chene(dbchene, idchene)
}
\arguments{
\item{dbchene}{dbchene TOCOMPLETE}

\item{idchene}{idchene TOCOMPLETE}
}
\value{
a list
}
\description{
Description
}
\examples{
conn <- DBI::dbConnect(RSQLite::SQLite(), system.file("chenes_truffe.sqlite", package = "truffles"))
chene <- DBI::dbReadTable(conn, name = "chenes_feularde")

get_info_chene(dbchene = chene, idchene = "chenes162")
DBI::dbDisconnect(conn)


}
