% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_chene.R
\name{get_info_chene}
\alias{get_info_chene}
\title{Get information about an oak tree}
\usage{
get_info_chene(dbchene, theidoak)
}
\arguments{
\item{dbchene}{Database containing information about oak trees.}

\item{theidoak}{ID of the oak tree for which information is to be retrieved.}
}
\value{
A list containing information about the oak tree, including its type and
date of plantation.

a list
}
\description{
This function retrieves information about a specific oak tree based on its ID
from the provided oak tree database.
}
\examples{

conn <- DBI::dbConnect(
   SQLite(),
   system.file(dbname, package = "truffles")
 )
chene <- DBI::dbReadTable(conn, name = "chenes_feularde")

get_info_chene(dbchene = chene, theidoak = "162")

}
