% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_chene_truffe.R
\name{get_info_chene_truffe}
\alias{get_info_chene_truffe}
\title{Get information about truffles associated with an oak tree}
\usage{
get_info_chene_truffe(dbtruffe, theidoak)
}
\arguments{
\item{dbtruffe}{Database containing information about truffles.}

\item{theidoak}{ID of the oak tree for which information about associated truffles
is to be retrieved.}
}
\value{
A list containing information about truffles associated with the oak tree,
including the total weight, date of the last truffle found, and any comments.
}
\description{
This function retrieves information about truffles associated with a specific oak tree
based on the oak tree's ID from the provided truffle database.
}
\examples{
conn <- DBI::dbConnect(
   RSQLite::SQLite(),
   system.file(dbname, package = "truffles")
 )
truffe <- DBI::dbReadTable(conn, name = "truffe")

get_info_chene_truffe(dbtruffe = truffe, theidoak = "119")
DBI::dbDisconnect(conn)
}
