% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_leaflet.R
\name{prepare_leaflet}
\alias{prepare_leaflet}
\title{Prepare Leaflet Data}
\usage{
prepare_leaflet(dbchene, dbtruffe)
}
\arguments{
\item{dbchene}{A data frame containing information about oak trees.}

\item{dbtruffe}{A data frame containing information about truffles.}
}
\value{
A list suitable for passing to Leaflet JavaScript for visualization.
}
\description{
This function prepares data for Leaflet JavaScript visualization.
}
\examples{
dbchene <- data.frame(
  id = c(1, 2, 3),
  chene_nom = c("Chene 1", "Chene 2", "Chene 3")
)
dbtruffe <- data.frame(
  idchene = c(1, 1, 3),
  idtruffe = c(1, 2, 3),
  estimation = c(1, 0, 1),
  poids = c(10, 15, NA)
)
prepare_leaflet(dbchene, dbtruffe)
}
