% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truffles_by_year.R, R/weight_truffles_by.R
\name{weight_truffles_by}
\alias{weight_truffles_by}
\title{Title}
\usage{
weight_truffles_by(truffles, ...)

weight_truffles_by(truffles, ...)
}
\arguments{
\item{truffles}{database}
}
\value{
a dataframe

a dataframe
}
\description{
Description

Description
}
\examples{
conn <- connect_db()
truffes <- DBI::dbReadTable(conn, name = "truffe")

weight_truffles_by(truffes, annee)

truffes_chene <- truffes |>
  inner_join(DBI::dbReadTable(conn, name = "chenes_feularde"), by = join_by(idchene == id))
weight_truffles_by(truffes_chene, annee, type)
conn <- connect_db()
truffes <- DBI::dbReadTable(conn, name = "truffe")

weight_truffles_by_year(truffes, annee)

truffes_chene <- truffes |>
  inner_join(DBI::dbReadTable(conn, name = "chenes_feularde"), by = join_by(idchene == id))
weight_truffles_by_year(truffes_chene, annee, type)
}
