% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_truffles_by.R
\name{weight_truffles_by}
\alias{weight_truffles_by}
\title{Calculate total weight of truffles grouped by specified columns}
\usage{
weight_truffles_by(dbtruffe, ...)
}
\arguments{
\item{dbtruffe}{A data frame containing information about truffles.}

\item{...}{Columns to group by, passed as arguments.}
}
\value{
A summarized data frame containing the total weight of truffles
grouped by the specified columns.
}
\description{
This function calculates the total weight of truffles grouped by specified columns
in a given data frame.
}
\examples{

library(dplyr)

conn <- DBI::dbConnect(
   SQLite(),
   system.file(dbname, package = "truffles")
 )
truffes <- DBI::dbReadTable(conn, name = "truffe")

weight_truffles_by(truffes, annee = lubridate::year(as.Date(date_found)))

truffes_chene <- truffes |>
  inner_join(DBI::dbReadTable(conn, name = "chenes_feularde"), by = "idoak")
weight_truffles_by(truffes_chene, annee = lubridate::year(as.Date(date_found)), type)
}
