% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_db_new_truffe.R
\name{write_db_new_truffe}
\alias{write_db_new_truffe}
\title{write_db_new_truffe}
\usage{
write_db_new_truffe(conn = connect_db(), id_chene, date_trouvee, poids)
}
\arguments{
\item{conn}{a connection to database}

\item{id_chene}{TOCOMPLETE}

\item{date_trouvee}{TOCOMPLETE}

\item{poids}{TOCOMPLETE}
}
\description{
Description
}
\examples{
# write_db_new_truffe(id_chene = "chene")
}
